requirement.txt
langchain==0.3.26
langchain-anthropic==0.3.16
langchain-aws==0.2.27
langchain-chroma==0.2.4
langchain-community==0.3.27
langchain-core==0.3.70
langchain-google-genai==2.1.8
langchain-google-vertexai==2.0.27
langchain-openai==0.3.27
langchain-tavily==0.2.7
langchain-text-splitters==0.3.8
langgraph==0.6.6
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.2
langgraph-supervisor==0.0.27
langgraph-swarm==0.0.12
neo4j==5.28.1
python-dotenv==1.0.1
typing_extensions==4.12.2
